C88-D1 refactor: fix fast refresh linting error.
