+++ src/components/YearSelect/YearSelect.stories.tsx
@@ -1,36 +1,31 @@
 import type { Meta, StoryObj } from '@storybook/react-vite'
-import { useState } from 'react'
 import YearSelect from './YearSelect'
+import { BrowserRouter } from 'react-router-dom'
+import { FilterSelectProvider } from '../../contexts/FilterSelectProvider'
 
 const meta: Meta<typeof YearSelect> = {
   title: 'Components/YearSelect',
   component: YearSelect,
   parameters: { layout: 'centered' },
+  decorators: [
+    (Story) => (
+      <BrowserRouter>
+        <FilterSelectProvider>
+          <Story />
+        </FilterSelectProvider>
+      </BrowserRouter>
+    ),
+  ],
 }
 
 type Story = StoryObj<typeof YearSelect>
 
-const StoryBookYearSelect = (props: React.ComponentProps<typeof YearSelect>) => {
-  const { selectedYear: initialSelectedYear = 2020, onChange, ...rest } = props
-  const [selectedYear, setSelectedYear] = useState<number>(initialSelectedYear)
-
-  const handleChange = (year: number) => {
-    setSelectedYear(year)
-    onChange?.(year)
-  }
-
-  return <YearSelect selectedYear={selectedYear} onChange={handleChange} {...rest} />
-}
-
 export const Primary: Story = {
-  render: (args) => <StoryBookYearSelect {...args} />,
-  args: { selectedYear: 2020 },
+  args: {},
 }
 
 export const Disabled: Story = {
-  render: (args) => <StoryBookYearSelect {...args} />,
   args: {
-    selectedYear: 2020,
     disabled: true,
   },
 }
