+++ src/main.tsx
@@ -7,7 +7,7 @@ import { StyledEngineProvider } from '@mui/material'
 import '../i18n'
 import MapContainer from './components/MapContainer/MapContainer'
 import NavigationHeader from './components/NavigationHeader/NavigationHeader'
-import { FilterSelectProvider } from './contexts/FilterSelectContext'
+import { FilterSelectProvider } from './contexts/FilterSelectProvider'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
